requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
click==8.1.2
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.10
django-rq==2.5.1
djangorestframework==3.13.1
packaging==21.3
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
rq==1.10.1
sqlparse==0.4.2
wrapt==1.14.0
